+++ crates/nostr-relay-builder/CHANGELOG.md
@@ -42,6 +42,7 @@
 - Add `LocalRelayBuilder::build` method (https://github.com/rust-nostr/nostr/pull/1147)
 - Impl `Default` for `LocalRelay` (https://github.com/rust-nostr/nostr/pull/1147)
 - Add `LocalRelay::sync_with` method (https://github.com/rust-nostr/nostr/pull/1146)
+- Reject expired events and ensure they are not sent to clients (https://github.com/rust-nostr/nostr/pull/1183)
 
 ## v0.44.0 - 2025/11/06